+++ packages/lib/uni-file-picker/uni-file-picker.vue
@@ -192,6 +192,13 @@ export default {
 					}
 					newFils.push(files ? files : v)
 				})
+				let data  = null
+				if (this.returnType === 'object') {
+					data = this.backObject(newFils)[0]
+				} else {
+					data = this.backObject(newFils)
+				}
+				this.formItem && this.formItem.setValue(data)
 				this.files = newFils
 			},
 			immediate: true
@@ -241,9 +248,29 @@ export default {
 	created() {
 		// this.files = Object.assign([], this.value)
 		this.tempData = {}
-
+		this.form = this.getForm('uniForms')
+		this.formItem = this.getForm('uniFormsItem')
+		if (this.form && this.formItem) {
+			if (this.formItem.name) {
+				this.rename = this.formItem.name
+				this.form.inputChildrens.push(this)
+			}
+		}
 	},
 	methods: {
+		/**
+		 * 获取父元素实例
+		 */
+		getForm(name = 'uniForms') {
+			let parent = this.$parent;
+			let parentName = parent.$options.name;
+			while (parentName !== name) {
+				parent = parent.$parent;
+				if (!parent) return false;
+				parentName = parent.$options.name;
+			}
+			return parent;
+		},
 		/**
 		 * 继续上传
 		 */
@@ -402,13 +429,14 @@ export default {
 					this.files[index].url = item.path
 					this.files[index].status = 'error'
 					this.files[index].errMsg = item.errMsg
-					this.files[index].progress = -1
+					// this.files[index].progress = -1
 					errorData.push(this.files[index])
 					errorTempFilePath.push(this.files[index].url)
 				} else {
 					this.files[index].errMsg = ''
 					this.files[index].url = item.url
 					this.files[index].status = 'success'
+					this.files[index].progress += 1
 					successData.push(this.files[index])
 					tempFilePath.push(this.files[index].url)
 				}
@@ -446,7 +474,8 @@ export default {
 				idx = this.files.findIndex(p => p.uuid === progressEvent.tempFile.uuid)
 			}
 			if (idx === -1 || !this.files[idx]) return
-			this.files[idx].progress = percentCompleted
+			// fix by mehaotian 100 就会消失，-1 是为了让进度条消失
+			this.files[idx].progress = percentCompleted - 1
 			// 上传中
 			this.$emit('progress', {
 				index: idx,
@@ -460,8 +489,10 @@ export default {
 		 * @param {Object} index
 		 */
 		delFile(index) {
-			let fileData = this.files[index]
-			this.$emit('delete', fileData)
+			this.$emit('delete', {
+				tempFile: this.files[index],
+				tempFilePath: this.files[index].url
+			})
 			this.files.splice(index, 1)
 			this.$nextTick(()=>{
 				this.setEmit()